requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
SQLAlchemy
databases
passlib
asyncpg
python-jose
passlib
python-multipart